+++ app/src/main/java/eu/kanade/tachiyomi/data/vault/operation/VaultChapterDeletePayload.kt
@@ -1,0 +1,15 @@
+package eu.kanade.tachiyomi.data.vault.operation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class VaultChapterDeletePayload(
+    val version: Int = VERSION,
+    val mangaId: Long,
+    val chapterId: Long,
+    val chapterTitle: String,
+) {
+    companion object {
+        const val VERSION = 1
+    }
+}