+++ app/src/main/java/eu/kanade/tachiyomi/data/vault/operation/VaultOperationHandler.kt
@@ -1,5 +1,6 @@
 package eu.kanade.tachiyomi.data.vault.operation
 
+import tachiyomi.domain.vault.model.VaultTransferJob
 import tachiyomi.domain.vault.model.VaultTransferState
 import tachiyomi.domain.vault.model.VaultTransferType
 
@@ -14,7 +15,7 @@ interface VaultOperationHandler {
     val type: VaultTransferType
     val policy: VaultOperationPolicy
 
-    suspend fun execute(payloadJson: String): VaultOperationExecutionResult
+    suspend fun execute(job: VaultTransferJob, payloadJson: String): VaultOperationExecutionResult
 }
 
 data class VaultOperationExecutionResult(