+++ app/src/main/java/eu/kanade/tachiyomi/di/AppModule.kt
@@ -29,12 +29,14 @@ import eu.kanade.tachiyomi.data.vault.localimport.LocalVaultChapterStager
 import eu.kanade.tachiyomi.data.vault.localimport.LocalVaultImportService
 import eu.kanade.tachiyomi.data.vault.localimport.LocalVaultMangaScanner
 import eu.kanade.tachiyomi.data.vault.localimport.LocalVaultMangaScannerBoundary
+import eu.kanade.tachiyomi.data.vault.operation.VaultChapterDeleteOperationHandler
 import eu.kanade.tachiyomi.data.vault.operation.VaultMetadataPublishOperationHandler
 import eu.kanade.tachiyomi.data.vault.operation.VaultOperationHandler
 import eu.kanade.tachiyomi.data.vault.operation.VaultOperationManager
 import eu.kanade.tachiyomi.data.vault.publishing.DefaultVaultChapterThumbnailCacheStore
 import eu.kanade.tachiyomi.data.vault.publishing.DefaultVaultChapterThumbnailDisplayLoader
 import eu.kanade.tachiyomi.data.vault.publishing.DefaultVaultChapterThumbnailPublishService
+import eu.kanade.tachiyomi.data.vault.publishing.VaultChapterDeletionService
 import eu.kanade.tachiyomi.data.vault.publishing.VaultChapterThumbnailCacheStore
 import eu.kanade.tachiyomi.data.vault.publishing.VaultChapterThumbnailDisplayLoader
 import eu.kanade.tachiyomi.data.vault.publishing.VaultChapterThumbnailPublishService
@@ -171,10 +173,19 @@ class AppModule(val app: Application) : InjektModule {
         addSingletonFactory {
             VaultMangaDeletionService(get<NetworkHelper>(), get(), get(), get(), get(), get(), get())
         }
+        addSingletonFactory {
+            VaultChapterDeletionService(get<NetworkHelper>(), get(), get(), get(), get(), get(), get())
+        }
         addSingletonFactory { VaultCoverPublishService(get<NetworkHelper>(), get(), get(), get(), get(), get()) }
         addSingletonFactory { VaultMetadataPublishService(get<NetworkHelper>(), get(), get(), get(), get()) }
-        addSingletonFactory<VaultOperationHandler> { VaultMetadataPublishOperationHandler(get(), get()) }
-        addSingletonFactory<List<VaultOperationHandler>> { listOf(get<VaultOperationHandler>()) }
+        addSingletonFactory { VaultMetadataPublishOperationHandler(get(), get()) }
+        addSingletonFactory { VaultChapterDeleteOperationHandler(get(), get()) }
+        addSingletonFactory<List<VaultOperationHandler>> {
+            listOf(
+                get<VaultMetadataPublishOperationHandler>(),
+                get<VaultChapterDeleteOperationHandler>(),
+            )
+        }
         addSingletonFactory { VaultOperationManager(app, get(), get()) }
         addSingletonFactory<VaultChapterThumbnailCacheStore> { DefaultVaultChapterThumbnailCacheStore(get()) }
         addSingletonFactory<VaultChapterThumbnailPublishService> {